+++ lock4j-core/src/main/java/com/baomidou/lock/DefaultLockFailureStrategy.java
@@ -18,7 +18,8 @@ package com.baomidou.lock;
 
 import com.baomidou.lock.exception.LockFailureException;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.util.StringUtils;
+
+import java.lang.reflect.Method;
 
 /**
  * @author zengzhihong
@@ -29,7 +30,7 @@ public class DefaultLockFailureStrategy implements LockFailureStrategy {
     protected static String DEFAULT_MESSAGE = "request failed,please retry it.";
 
     @Override
-    public void onLockFailure(String key, String message) {
-        throw new LockFailureException(!StringUtils.hasText(message) ? DEFAULT_MESSAGE : message);
+    public void onLockFailure(String key, Method method, Object[] arguments) {
+        throw new LockFailureException(DEFAULT_MESSAGE);
     }
 }