+++ lock4j-core/src/main/java/com/baomidou/lock/aop/LockInterceptor.java
@@ -19,7 +19,6 @@ package com.baomidou.lock.aop;
 import com.baomidou.lock.LockFailureStrategy;
 import com.baomidou.lock.LockInfo;
 import com.baomidou.lock.LockKeyBuilder;
-import com.baomidou.lock.LockMessageBuilder;
 import com.baomidou.lock.LockTemplate;
 import com.baomidou.lock.annotation.Lock4j;
 import com.baomidou.lock.spring.boot.autoconfigure.Lock4jProperties;
@@ -27,6 +26,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.aopalliance.intercept.MethodInterceptor;
 import org.aopalliance.intercept.MethodInvocation;
+import org.springframework.util.StringUtils;
 
 
 /**
@@ -42,8 +42,6 @@ public class LockInterceptor implements MethodInterceptor {
 
     private final LockKeyBuilder lockKeyBuilder;
 
-    private final LockMessageBuilder lockMessageBuilder;
-
     private final LockFailureStrategy lockFailureStrategy;
 
     private final Lock4jProperties lock4jProperties;
@@ -54,22 +52,15 @@ public class LockInterceptor implements MethodInterceptor {
         try {
             Lock4j lock4j = invocation.getMethod().getAnnotation(Lock4j.class);
             String prefix = lock4jProperties.getLockKeyPrefix() + ":";
-            switch (lock4j.scope()) {
-                case GLOBAL:
-                    break;
-                case METHOD:
-                    prefix += invocation.getMethod().getDeclaringClass().getName() + invocation.getMethod().getName();
-                    break;
-                default:
-            }
+            prefix += StringUtils.hasText(lock4j.name()) ? lock4j.name() :
+                    invocation.getMethod().getDeclaringClass().getName() + invocation.getMethod().getName();
             String key = prefix + "#" + lockKeyBuilder.buildKey(invocation, lock4j.keys());
             lockInfo = lockTemplate.lock(key, lock4j.expire(), lock4j.acquireTimeout(), lock4j.executor());
             if (null != lockInfo) {
                 return invocation.proceed();
             }
             // lock failure
-            lockFailureStrategy.onLockFailure(key, lockMessageBuilder.buildMessage(invocation,
-                    lock4j.message()));
+            lockFailureStrategy.onLockFailure(key, invocation.getMethod(), invocation.getArguments());
             return null;
         } finally {
             if (null != lockInfo) {